DirectX12 Barrier support beginning of support. Can now render to target and use as texture. Can update gpu data (vertices, textures, ...) All preliminary, needs lot of cleanup Crash after a short while

+++ Engine/ZenApi/ZenCore/Constant/zenCoreConstResource.h
@@ -166,7 +166,7 @@ namespace zen {
 	namespace zenType
 	{
 		typedef zFlag<eResType,		keResType__Count, zU64>	zFlagResType;
-		typedef zFlag<ePlatform,	(int)kePlatform__Count>		zFlagResPlatform;
+		typedef zFlag<ePlatform,	kePlatform__Count>		zFlagResPlatform;
 		typedef zFlag<eResourceUse,	keResUse__Count>		zFlagResUse;
 		typedef zFlag<eTextureUse,	keTexCreate__Count>		zFlagResTexUse;		
 	}
